backend/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
django-environ==0.12.0
psycopg[binary]==3.3.2
celery==5.6.2
redis==5.2.1
whitenoise==6.11.0
pyotp==2.9.0
qrcode==7.3
pillow==11.3.0
reportlab==4.4.5
cryptography==45.0.6
gunicorn==25.0.3
openpyxl==3.1.5
certifi==2025.1.31
python-docx==1.1.2
